+++ pyengine/Window.py
@@ -1,6 +1,5 @@
 import pygame
-from pyengine.Exceptions import NoObjectError
-from pyengine.GameState import GameState
+from pyengine.World import World
 from pyengine.Utils import Color, Colors
 from pygame import locals as const
 
@@ -18,8 +17,7 @@ class Window:
         self.clock = pygame.time.Clock()
         self.width = width
         self.height = height
-        self.states = []
-        self.current_state = None
+        self.world = World(self)
         self.launch = True
         self.debug = debug
         self.color = color
@@ -52,45 +50,28 @@ class Window:
     def set_debug(self, debug):
         self.debug = debug
 
-    def add_state(self, state):
-        if not isinstance(state, GameState):
-            raise TypeError("Argument is not type of "+str(GameState)+" but "+str(type(state))+".")
-        if len(self.states) == 0:
-            self.current_state = state
-        state.set_window(self)
-        self.states.append(state)
+    def set_world(self, world):
+        self.world = world
 
-    def set_current_state(self, name):
-        for i in self.states:
-            if i.name == name:
-                self.current_state = i
-
-    def get_current_state(self):
-        return self.current_state
-
-    def get_state(self, name):
-        for i in self.states:
-            if i.name == name:
-                return i
+    def get_world(self):
+        return self.world
 
     def process_event(self, evt):
         if evt.type == const.QUIT:
             self.launch = False
         elif evt.type == const.KEYDOWN:
-            self.current_state.keypress(evt)
+            self.world.keypress(evt)
         elif evt.type == const.MOUSEBUTTONDOWN:
-            self.current_state.mousepress(evt)
+            self.world.mousepress(evt)
         elif evt.type == const.KEYUP:
-            self.current_state.keyup(evt)
+            self.world.keyup(evt)
         else:
-            self.current_state.event(evt)
+            self.world.event(evt)
 
     def stop(self):
         self.launch = False
 
     def run(self):
-        if len(self.states) == 0:
-            raise NoObjectError("Window have any GameState")
         while self.launch:
             for event in pygame.event.get():
                 self.process_event(event)
@@ -98,7 +79,7 @@ class Window:
             self.screen.fill(self.color.get())
             self.clock.tick(60)
 
-            self.current_state.run()
+            self.world.run()
 
             pygame.display.update()
         pygame.quit()