+++ pyengine/__init__.py
@@ -1,7 +1,7 @@
 try:
     from pyengine.Window import Window
     from pyengine.Entity import Entity
-    from pyengine.GameState import GameState, StateCallbacks
+    from pyengine.World import World, WorldCallbacks
     from pyengine.Components.PhysicsComponent import CollisionCauses
     from pyengine.Components.ControlComponent import ControlType, Controls, MouseButton
     from pygame import locals as const
@@ -9,5 +9,5 @@ except ModuleNotFoundError:
     pass
 
 __all__ = ["Window", "Entity", "ControlType", "const", "MouseButton", "CollisionCauses",
-           "StateCallbacks", "GameState", "Controls"]
-__version__ = "1.1.2"
+           "WorldCallbacks", "World", "Controls"]
+__version__ = "1.2.0"